requirements.txt
mysql-connector-python==8.0.28
Flask==2.1.2
pandas==1.4.0
scikit-learn==1.0.2
nltk==3.6.5
